PriorityGrid.
PriorityGrid is created in main.cpp
When an app is added to the rMonitor, the pGrid is passed to the
PriorityData of the app.

App adds its Pvisual and Pinteract to the underlying cells.
This is done in (SN_PriorityData::computePriority())

A complete priorityGrid can only be obtained at every scheduling
iteration.